requirements.txt
requests>=2.31.0
python-dotenv>=1.0.1
Pillow>=10.4.0
pyinstaller>=6.10.0
